# Backend_ASP.NET/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Supplies API",
    "version": "v1"
  },
  "paths": {
...
    "/Admin/User/Edit/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          },
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "PassWord",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Roles",
            "type": "string"
          },
          {
            "in": "query",
            "name": "UserName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "FirstName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "LastName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PhoneNumber",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Avata",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/User/Delete/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/User/Create": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "FirstName",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "LastName",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "Email",
            "required": true,
            "type": "string",
            "format": "email"
          },
          {
            "in": "query",
            "name": "PhoneNumber",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "Password",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "ConfirmPassword",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Admin/User/Login": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "PhoneNumber",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "Password",
            "required": true,
            "type": "string",
            "format": "password"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Customers/GetByID/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Customers/Update/{id}": {
      "put": {
        "tags": [
          "Customers"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CustomerModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Customers/Delete/{id}": {
      "delete": {
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Customers/Create": {
      "post": {
        "tags": [
          "Customers"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "storeID",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CustomerModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Debits/GetByID/{id}": {
      "get": {
        "tags": [
          "Debits"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Debits/Update/{id}": {
      "put": {
        "tags": [
          "Debits"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DebitsModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Debits/Delete/{id}": {
      "delete": {
        "tags": [
          "Debits"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Debits/Create": {
      "post": {
        "tags": [
          "Debits"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DebitsModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CustomerModel": {
      "type": "object",
      "properties": {
        "custommerId": {
          "format": "uuid",
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        },
        "sex": {
          "$ref": "#/definitions/sexEnum"
        },
        "address": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DebitsModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "note": {
          "type": "string"
        },
        "totalMoney": {
          "format": "int32",
          "type": "integer"
        },
        "paymentStatus": {
          "type": "boolean"
        },
        "debPurchaseDate": {
          "format": "date-time",
          "type": "string"
        },
        "customerID": {
          "format": "uuid",
          "type": "string"
        },
        "employeeID": {
          "format": "uuid",
          "type": "string"
        },
        "storeID": {
          "format": "uuid",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "sexEnum": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    }
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}